api/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.12
pillow>=10.4.0
httpx>=0.27.2
